+++ lib/report.mli
@@ -36,7 +36,10 @@ end
 val empty : unit -> t
 
 val of_krs :
-  ?okr_db:Masterdb.t -> KR.t list -> t * [> KR.Warning.t | KR.Error.t ] list
+  ?okr_db:Masterdb.t ->
+  ?week:Week.t ->
+  KR.t list ->
+  t * [> KR.Warning.t | KR.Error.t ] list
 
 val of_markdown :
   ?existing_report:t ->
@@ -44,6 +47,7 @@ val of_markdown :
   ?include_sections:string list ->
   ?okr_db:Masterdb.t ->
   ?report_kind:Parser.report_kind ->
+  ?week:Week.t ->
   Parser.markdown ->
   t * [> KR.Warning.t | KR.Error.t ] list
 
@@ -57,7 +61,11 @@ val iter :
 val find : t -> KR.Id.t -> KR.t list
 
 val add :
-  ?okr_db:Masterdb.t -> t -> KR.t -> [> KR.Warning.t | KR.Error.t ] option
+  ?okr_db:Masterdb.t ->
+  ?week:Week.t ->
+  t ->
+  KR.t ->
+  [> KR.Warning.t | KR.Error.t ] option
 
 val all_krs : t -> KR.t list