+++ CHANGES.md
@@ -1,3 +1,9 @@
+## unreleased
+
+### Changed
+
+- Lint: using a work-item instead of an objective raises an error starting from Week 24 (June 10, 2024) (#270, @gpetiot)
+
 ## 2.1.0
 
 ### Changed